Consolidated manip.go and km.go.
